+++ container-engine-lib/lib/backend_impls/metrics_reporting/metrics_reporting_kurtosis_backend.go
@@ -10,7 +10,6 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/networking_sidecar"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/shell"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/wait_for_availability_http_methods"
 	"github.com/kurtosis-tech/stacktrace"
 	"io"
@@ -410,19 +409,19 @@ func (backend *MetricsReportingKurtosisBackend) WaitForUserServiceHttpEndpointAv
 	return nil
 }
 
-func (backend *MetricsReportingKurtosisBackend) GetShellOnUserService(
+func (backend *MetricsReportingKurtosisBackend) GetConnectionWithUserService(
 	ctx context.Context,
 	enclaveId enclave.EnclaveID,
 	serviceGUID service.ServiceGUID,
 )(
-	resultShell *shell.Shell,
+	resultConn net.Conn,
 	resultErr error,
 ) {
-	newShell, err := backend.underlying.GetShellOnUserService(ctx, enclaveId, serviceGUID)
+	newConn, err := backend.underlying.GetConnectionWithUserService(ctx, enclaveId, serviceGUID)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred getting shell on user service with GUID '%v'", serviceGUID)
+		return nil, stacktrace.Propagate(err, "An error occurred getting connection with user service with GUID '%v'", serviceGUID)
 	}
-	return newShell, nil
+	return newConn, nil
 }
 
 func (backend *MetricsReportingKurtosisBackend) StopUserServices(