+++ container-engine-lib/lib/backend_interface/kurtosis_backend.go
@@ -9,7 +9,6 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/networking_sidecar"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/shell"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/wait_for_availability_http_methods"
 	"io"
 	"net"
@@ -255,13 +254,13 @@ type KurtosisBackend interface {
 		resultErr error,
 	)
 
-	// Get an interactive shell to execute commands in an user service
-	GetShellOnUserService(
+	// Get a connection with user service to execute commands in
+	GetConnectionWithUserService(
 		ctx context.Context,
 		enclaveId enclave.EnclaveID,
 		serviceGUID service.ServiceGUID,
 	)(
-		resultShell *shell.Shell,
+		resultConn net.Conn,
 		resultErr error,
 	)